+++ backend/beets_flask/server/routes/config.py
@@ -4,7 +4,6 @@ we need some of our settings in the frontend. proposed solution:
 fetch settings once from the backend on first page-load.
 """
 
-from beets import __version__ as beets_version
 from quart import Blueprint, jsonify
 
 from beets_flask.config import get_config
@@ -16,46 +15,20 @@ config_bp = Blueprint("config", __name__, url_prefix="/config")
 async def get_all():
     """Get nested dict representing the full (but redacted) beets config."""
     config = get_config()
-    return jsonify(_serializable(config.flatten(redact=True)))
+    return jsonify(_serializable(config.beets_config.flatten(redact=True)))
 
 
 @config_bp.route("/", methods=["GET"])
 async def get_basic():
     """Get the config settings needed for the gui."""
     config = get_config()
-    plugins = config["plugins"].as_str_seq()
-    from beets.metadata_plugins import find_metadata_source_plugins
-
-    data_sources: list[str] = [
-        p.__class__.data_source for p in find_metadata_source_plugins()
-    ]
 
     return jsonify(
         {
-            "gui": _serializable(config["gui"].flatten(redact=True)),
-            "import": {
-                k: config["import"][k].get()
-                for k in [
-                    "duplicate_action",
-                ]
-            },
-            "match": {
-                k: config["match"][k].get()
-                for k in [
-                    "strong_rec_thresh",
-                    "medium_rec_thresh",
-                ]
-            }
-            | {
-                k: config["match"][k].as_str_seq()
-                for k in [
-                    "album_disambig_fields",
-                    "singleton_disambig_fields",
-                ]
-            },
-            "plugins": config["plugins"].as_str_seq(),
-            "data_sources": data_sources,
-            "beets_version": beets_version,
+            **config.to_dict(extra_fields=False),
+            # the following ones are not part of the schema
+            "beets_metadata_sources": config.beets_metadata_sources,
+            "beets_version": config.beets_version,
         }
     )
 
@@ -91,9 +64,10 @@ async def refresh():
     curl -X POST http://localhost:5001/api_v1/config/refresh
     ```
     """
-    from beets_flask.config.beets_config import refresh_config
+    from beets_flask.config import get_config
 
-    refresh_config()
+    config = get_config()
+    config.reload()
     return jsonify({"status": "ok"})