+++ backend/tests/integration/test_flows.py
@@ -15,6 +15,7 @@ import pytest
 from sqlalchemy import delete, func, select
 from sqlalchemy.orm import Session
 
+from beets_flask.config.beets_config import get_config
 from beets_flask.database.models.states import (
     FolderInDb,
     SessionStateInDb,
@@ -894,6 +895,13 @@ class TestMultipleTasks(
         path_multiple_tasks: Path,
     ):
         """Test the import of the tagged folder."""
+
+        # avoid duplicate errors when re-importing.
+        # TODO: fix beets lib mixin, this should not be necessary,
+        # if library is cleared correctly.
+        config = get_config()
+        config.data.import_.duplicate_action = "remove"
+        config.commit_to_beets()
 
         exc = await run_preview(
             "obsolete_hash_preview",